cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "securelens-ai"
version = "2.0.0"
description = "AI-powered CLI security scanner for codebases and URLs"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["security", "cli", "sast", "ai", "scanner", "owasp"]
dependencies = [
    "click>=8.1",
    "rich>=13.7",
    "litellm>=1.40",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "questionary>=2.0",
    "pathspec>=0.12",
    "asyncio-throttle>=1.0",
]

[project.scripts]
securelens = "securelens.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["securelens*"]
